requirements.txt
Flask
Flask-Cors
PyJWT
pymongo
python-dotenv
Flask-Mail
stripe
pyngrok